services/traffic-gen/requirements.txt
httpx==0.27.0
kafka-python==2.0.2